+++ R/multi_scale.R
@@ -365,6 +365,13 @@ ww_multi_scale.sf <- function(
     )
   }
 
+  if (any(names(data) %in% c(".truth", ".estimate", ".truth_count", ".estimate_count"))) {
+    rlang::abort(c(
+      "This function cannot work with data whose columns are named `.truth`, `.estimate`, `.truth_count`, or `estimate_count`.",
+      i = "Rename the relevant columns and try again."
+    ))
+  }
+
   geom_type <- unique(sf::st_geometry_type(data))
   if (!(length(geom_type) == 1 && geom_type == "POINT")) {
     rlang::abort(
@@ -447,9 +454,9 @@ ww_multi_scale.sf <- function(
       matched_data <- dplyr::summarise(
         matched_data,
         .truth = rlang::exec(.env[["aggregation_function"]], .data[[names(truth_var)]]),
-        .truth_count = sum(!is.na({{ truth }})),
+        .truth_count = sum(!is.na(.data[[names(truth_var)]])),
         .estimate = rlang::exec(.env[["aggregation_function"]], .data[[names(estimate_var)]]),
-        .estimate_count = sum(!is.na({{ estimate }})),
+        .estimate_count = sum(!is.na(.data[[names(estimate_var)]])),
         .groups = "drop"
       )
 
@@ -522,9 +529,16 @@ handle_grids <- function(data, grids, autoexpand_grid, ...) {
     grids <- lapply(
       grid_arg_idx,
       function(idx) {
+        arg <- lapply(
+          names(grid_args),
+          function(arg) {
+            grid_args[[arg]][[idx]]
+          }
+        )
+        names(arg) <- names(grid_args)
         do.call(
           sf::st_make_grid,
-          c(x = list(grid_box), grid_args[idx, ])
+          c(x = list(grid_box), arg)
         )
       }
     )