+++ tests/testthat/test-multi_scale.R
@@ -571,3 +571,126 @@ test_that("units are handled properly", {
   )
 
 })
+
+test_that("counts of non-NA values are correct", {
+  pts <- sf::st_sample(
+    sf::st_as_sfc(
+      sf::st_bbox(
+        c(xmin = 1327326, ymin = 2175524, xmax = 1971106, ymax = 2651347),
+        crs = 5072
+      )
+    ),
+    500
+  )
+
+  pts <- sf::st_as_sf(pts)
+  pts$truth <- rnorm(500, 123, 35)
+  pts$estimate <- rnorm(500, 123, 39)
+
+  cellsizes <- units::set_units(20, "km")
+
+  ww_output <- ww_multi_scale(
+    pts,
+    truth,
+    estimate,
+    cellsize = cellsizes,
+    square = FALSE,
+    metrics = yardstick::rmse
+  )
+
+  actual <- ww_output$.grid[[1]]$.truth_count
+  expected <- vapply(
+    seq_len(nrow(ww_output$.grid[[1]])),
+    function(idx) nrow(sf::st_intersection((ww_output$.grid[[1]][idx, ]["x"]), pts["x"])),
+    integer(1)
+  )
+  expected[expected == 0] <- NA_integer_
+  expect_identical(actual, expected)
+})
+
+test_that("counts are the same whether or not column names are quoted", {
+  pts <- sf::st_sample(
+    sf::st_as_sfc(
+      sf::st_bbox(
+        c(xmin = 1327326, ymin = 2175524, xmax = 1971106, ymax = 2651347),
+        crs = 5072
+      )
+    ),
+    500
+  )
+
+  pts <- sf::st_as_sf(pts)
+  pts$truth <- rnorm(500, 123, 35)
+  pts$estimate <- rnorm(500, 123, 39)
+
+  cellsizes <- units::set_units(seq(20, 100, 10), "km")
+
+  expect_identical(
+    lapply(
+      waywiser::ww_multi_scale(
+        pts,
+        truth,
+        estimate,
+        cellsize = cellsizes,
+        square = FALSE,
+        metrics = yardstick::rmse
+      )$.grid,
+      function(x) x$.truth_count
+    ),
+    lapply(
+      waywiser::ww_multi_scale(
+        pts,
+        "truth",
+        "estimate",
+        cellsize = cellsizes,
+        square = FALSE,
+        metrics = yardstick::rmse
+      )$.grid,
+      function(x) x$.truth_count
+    )
+  )
+})
+
+test_that("using protected names triggers errors", {
+  skip_if_not_installed("units")
+  pts <- sf::st_sample(
+    sf::st_as_sfc(
+      sf::st_bbox(
+        c(xmin = 1327326, ymin = 2175524, xmax = 1971106, ymax = 2651347),
+        crs = 5072
+      )
+    ),
+    500
+  )
+
+  pts <- sf::st_as_sf(pts)
+  pts$.truth <- rnorm(500, 123, 35)
+  pts$.estimate <- rnorm(500, 123, 39)
+
+  cellsizes <- units::set_units(20, "km")
+
+  expect_snapshot_error(
+    ww_multi_scale(
+      pts,
+      .truth,
+      .estimate,
+      cellsize = cellsizes,
+      square = FALSE,
+      metrics = yardstick::rmse
+    )
+  )
+
+  pts$truth <- pts$.truth
+  pts$.truth <- NULL
+
+  expect_snapshot_error(
+    ww_multi_scale(
+      pts,
+      truth,
+      .estimate,
+      cellsize = cellsizes,
+      square = FALSE,
+      metrics = yardstick::rmse
+    )
+  )
+})